requirements.txt
python-telegram-bot==20.3
pydrive2==1.19.0
pycryptodome==3.20.0
requests==2.31.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
tqdm==4.66.1
pySmartDL==1.3.4
URLObject>=2.1.1
